+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### General
 
 - Throw `IndexError` when using negative indexes on `PaginatedList` objects (Thanks, [@UniversalSuperBox](https://github.com/UniversalSuperBox))
+- `Assignment.overrides` now returns a list of `AssignmentOverride` objects.
 
 ### Deprecation Warnings